requirements.txt
Django>=4.2.0,<5.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0
django-crispy-bootstrap5>=2023.1
django-allauth>=0.54.0
django-filter>=23.1
django-debug-toolbar>=4.1.0
django-extensions>=3.2.3
django-cleanup>=7.0.0
whitenoise>=6.5.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.6
python-dateutil>=2.8.2
pytz>=2023.3
sqlparse>=0.4.4
asgiref>=3.6.0
django-widget-tweaks>=1.5.0
